Added new cards to use
These cards are LGPL licensed and look pretty good. Instead of using
separate sprite sheets for each suit, I used a single sprite sheet with
a single image for the card back.

This required a work around a bit of technical debt. The code is
currently set up to use CardSprite's constructor when cards are being
created in either the hand or the discard pile. However, that was only
actually working for the hand cards, not the discard pile cards. This
was due to a simple issue.

CardGroupings, which hand and discard pile inherit from, overrides the
classType attribute to use the CardSprite constructor. However, so does
Object.Group, which CardGroupings inherits from. So CardGrouping was
essentially setting it, and the Object.Group was setting it back. So, I
just set CardGrouping to set it after the Object.Group constructor call
which seeemd to work.

Some logic needed to be reworked to handle a single sprite sheet, but
that wasn't too complicated.

I deleted some files which are no longer being used now.